Removed unnecessary casts

+++ nImO/nImOarray.cpp
@@ -296,7 +296,7 @@ nImO::Array::extractValue(const nImO::Message &theMessage,
             if (ReadStatus::Successful == status)
             {
                 ODL_LOG("(ReadStatus::Successful == status)"); //####
-                elementCount -= static_cast<int64_t>(DataKindIntegerShortValueMinValue) - 1;
+                elementCount -= DataKindIntegerShortValueMinValue - 1;
                 ODL_LL1("elementCount <- ", elementCount); //####
                 if (0 >= elementCount)
                 {
@@ -674,8 +674,7 @@ const
         std::queue<double> doublesSeen;
 
         outMessage.appendBytes(&startArray, sizeof(startArray));
-        writeInt64ToMessage(outMessage, inherited2::size() +
-                            static_cast<int64_t>(DataKindIntegerShortValueMinValue) - 1);
+        writeInt64ToMessage(outMessage, inherited2::size() + DataKindIntegerShortValueMinValue - 1);
         for (const_iterator walker(inherited2::begin()); (inherited2::end() != walker); ++walker)
         {
             SpValue aValue(*walker);
