+++ nImO/nImOset.cpp
@@ -299,7 +299,7 @@ nImO::Set::extractValue(const nImO::Message &theMessage,
             if (ReadStatus::Successful == status)
             {
                 ODL_LOG("(ReadStatus::Successful == status)"); //####
-                elementCount -= static_cast<int64_t>(DataKindIntegerShortValueMinValue) - 1;
+                elementCount -= DataKindIntegerShortValueMinValue - 1;
                 ODL_LL1("elementCount <- ", elementCount); //####
                 if (0 >= elementCount)
                 {
@@ -745,8 +745,7 @@ const
                            static_cast<uint8_t>(DataKind::OtherContainerNonEmptyValue);
 
         outMessage.appendBytes(&startSet, sizeof(startSet));
-        writeInt64ToMessage(outMessage, inherited2::size() +
-                            static_cast<int64_t>(DataKindIntegerShortValueMinValue) - 1);
+        writeInt64ToMessage(outMessage, inherited2::size() + DataKindIntegerShortValueMinValue - 1);
         for (const_iterator walker(inherited2::begin()); (inherited2::end() != walker); ++walker)
         {
             SpValue aValue(*walker);
